Add perfect forwarding to combine_hash and optimize combine_hash
Combine hash now links to a separate function named combine_hash_hash,
which sets the state to zero before actually starting the hash.
Beforehand, it would just use the first hash as the state. But it came
into an issue in that slight changes to the first hash wouldn't create
as much as an avalanche effect as the second, or even last element. This
helps make sure that all hashes are actually used

Thinking about it, all I really want in this is boost::combine_hash in
a way that allows me to enter in as many parameters as I need
But I don't want to use non public domain code in this project, and am
stuck implementing my own algorithm. From my limited testing, this works
slightly better than my previous one, as a slight change to the first
element won't have as much as an effect